requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-crontab==0.7.1
gunicorn==20.1.0
idna==3.3
-e git+https://github.com/iqoptionapi/iqoptionapi.git@8a903cc094a74af1ed935a56a2d6b5a9ed3319d7#egg=iqoptionapi
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mysql-connector-python==8.0.28
platformdirs==2.5.1
protobuf==3.19.4
psycopg2==2.9.3
pylint==2.12.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
websocket-client==0.56.0
whitenoise==6.0.0
wrapt==1.13.3
